Report and calibrate power with 0.01 W resolution (A1Z + BSEED)
Power was whole-watt end to end because the standard activePower attribute
(0x050B) is int16 and overflows in centiwatts above 327 W. Widen the power
path to int32 centiwatts and expose it on a new custom int32 attribute
0xFF13; the standard int16 activePower stays whole-watt for compatibility.

- energy_meter: power is now int32 cW; add overflow-safe 32-bit helpers
  energy_meter_product_to_cw() and energy_meter_power_mult_from_cw(). The
  calibration multiplier keeps its whole-watt meaning, so existing stored
  calibration (e.g. the BSEED outlets) stays valid.
- HLW8012 and BL0942 drivers output centiwatts; BL0942 energy integration
  uses whole watts to keep the accumulator's headroom. Both accept a
  centiwatt calibration reference.
- electrical measurement cluster: register 0xFF13 (int32 cW) and widen
  calibrate_power (0xFF12) to uint32; populate both power attributes.
- converter: power reads 0xFF13 with divisor 100 / precision 2; power is
  reported via 0xFF13 (reportableChange 50 = 0.5 W); calibrate_power takes
  W with 0.01 steps and scales x100 into the uint32 attribute.

Applies to every metering device (BL0942 EB and HLW8012 EP).

+++ src/base_components/energy_measurement/bl0942.h
@@ -16,7 +16,7 @@
 // divider; reference constants: 15883.34 counts/V, 251065.7 counts/A,
 // 623.03 counts/W). Physical value = raw * MULTIPLIER / FIXED_POINT_SCALE with
 // the same output units as the HLW8012 driver: voltage in cV, current in mA,
-// power in W. Field-adjustable via the on-device calibrate attributes or the
+// power in cW. Field-adjustable via the on-device calibrate attributes or the
 // config_str V/A/W markers.
 #define BL0942_FIXED_POINT_SCALE     65536
 #ifndef BL0942_VOLTAGE_MULTIPLIER
@@ -41,7 +41,7 @@
 typedef struct {
     uint16_t voltage;    // cV
     uint16_t current;    // mA
-    int16_t  power;      // W
+    int32_t  power;      // cW (centiwatts)
     uint32_t energy;     // Wh
     uint32_t energy_acc; // W*ms remainder (< 1 Wh)
     uint32_t last_frame_time;
